Update 2022-07-20-README.md
Looks like there was thin clouds that moved through at times reducing the brightness of all of the stars. Will have to decided which frames can be used for data. Also as a note, all images were taken for one star and then for the next star. Each target was taken in blocks instead of 3 images per target repeating.